requirements.txt
annotated-types==0.5.0
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dnspython==2.4.1
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
loguru==0.7.2
lxml==4.9.3
packaging==23.1
pluggy==1.2.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
scrapingbee==1.2.0
sniffio==1.3.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
win32-setctime==1.1.0
